fixes SpellCorrector
